Refactor lease duration from int32 to int and update lease acquisition logic

+++ internal/config/config.go
@@ -30,7 +30,7 @@ type Config struct {
 	// ReleaseOnExit releases the IP address on exit
 	ReleaseOnExit bool `json:"release-on-exit"`
 	// LeaseDuration is the duration of the kubernetes lease
-	LeaseDuration int32 `json:"lease-duration"`
+	LeaseDuration int `json:"lease-duration"`
 }
 
 func NewConfig(c *cli.Context) *Config {
@@ -46,6 +46,6 @@ func NewConfig(c *cli.Context) *Config {
 	cfg.Region = c.String("region")
 	cfg.IPv6 = c.Bool("ipv6")
 	cfg.ReleaseOnExit = c.Bool("release-on-exit")
-	cfg.LeaseDuration = int32(c.Int("lease-duration"))
+	cfg.LeaseDuration = c.Int("lease-duration")
 	return &cfg
 }
